fitness-tracker/requirements.txt
django==3.2
dj-database-url==1.2.0
gunicorn==20.1.0
tzdata==2022.7
pillow~=10.3.0
django-environ==0.9.0
djangorestframework-simplejwt~=5.2.2
django-rest-swagger~=2.2.0
django-cors-headers~=3.14.0
django-admin-interface~=0.24.2
django-smart-selects==1.5.9
djangorestframework~=3.15.1
django-environ~=0.9.0
requests~=2.28.1
drf-yasg~=1.21.7
